Move cmd api main() to direktiv main().
