+++ .gitignore
@@ -60,7 +60,6 @@ issuer.*
 !*build/docker/flow
 !*build/docker/sidecar
 !*build/docker/api
-!*build/docker/apps
 !*build/docker/functions
 /build/docker/sidecar/sidecar
 *build/docker/functions/functions
